Updated import query to clear existing relations and link to MONDO
We encountered an error when the structure of MONDO changed to list
OMIM terms separately and link to them with an equivalentTo
relation. This resulted in gene-dosage curations either being linked
to both MONDO and OMIM terms simultaneously or just to OMIM terms,
losing the information about the disease. The code was updated to
clear existing relationships and link specifically against MONDO terms.